Pr69v2

+++ NativeYoutube/ViewModels/AppStateViewModel.swift
@@ -11,33 +11,16 @@ import SwiftUI
 class AppStateViewModel: ObservableObject {
     // MARK: Global values
 
-    @AppStorage(AppStorageStrings.apiKey.rawValue) var apiKey = Constants.defaultAPIKey
-    @AppStorage(AppStorageStrings.playListID.rawValue) var playListID = Constants.defaultPlaylistID
-    @AppStorage(AppStorageStrings.useIINA.rawValue) var useIINA: Bool = false
+    @AppStorage(AppStorageStrings.apiKey) var apiKey = Constants.defaultAPIKey
+    @AppStorage(AppStorageStrings.playListID) var playListID = Constants.defaultPlaylistID
+    @AppStorage(AppStorageStrings.useIINA) var useIINA: Bool = false
+    @AppStorage(AppStorageStrings.videoClickBehaviour) var vidClickBehaviour: VideoClickBehaviour = .playVideo
 
     // MARK: States
 
     @Published var logs: [String] = .init()
     @Published var isPlaying: Bool = false
     @Published var currentlyPlaying: String = ""
-
-    func changePlayListID(for playlistURL: String) -> Bool {
-        //        Using regex would be 100% better (this is a quick and dirty method)
-        var changed = false
-        let splitted = playlistURL.split(separator: "&")
-        for splitted in splitted {
-            if splitted.contains("list") {
-                let id = splitted.split(separator: "=")
-                if let idCount = id.last?.count {
-                    if idCount > 6 {
-                        playListID = String(id.last!)
-                        changed = true
-                    }
-                }
-            }
-        }
-        return changed
-    }
 
     func addToLogs(for page: Pages, message: String) {
         logs.append("Log at: \(Date()), from \(page.rawValue), message => \(message)")
@@ -46,10 +29,12 @@ class AppStateViewModel: ObservableObject {
     func stopPlaying() {
         currentlyPlaying = ""
         isPlaying = false
-        Task {
-            let shellOutput = shell("killall IINA")
-            DispatchQueue.main.async {
-                self.logs.append(shellOutput)
+        if useIINA {
+            Task {
+                let shellOutput = shell("killall IINA")
+                DispatchQueue.main.async {
+                    self.logs.append(shellOutput)
+                }
             }
         }
     }
@@ -61,7 +46,7 @@ class AppStateViewModel: ObservableObject {
     }
 
     func playVideoIINA(url: URL, title: String) {
-//        togglePlaying(title)
+        togglePlaying(title)
         Task {
             DispatchQueue.main.async {
                 self.logs.append(self.shell("open -a iina '\(url)'"))
